[FE] refactor: 폴더 생성
